==> notebooks/01_api_get_spain_leagues.ipynb <==
import os
os.chdir('C://Users//T420//Desktop//UBIQUM//5 PROJECT//Quiniela//notebooks')
os.getcwd()

import requests
url = "https://server1.api-football.com/leagues/country/Spain"
headers = {'x-rapidapi-key': "28d61524ab5698a1f8edb9e9ed577221"}

response = requests.request("GET", url, headers=headers)

spain_leagues = response.json()
spain_leagues

league_id = []
season =[]
season_start = []
season_end = []
name = []

for i in spain_leagues['api']['leagues']:
    league_id.append(i['league_id'])
    season.append(i['season'])
    season_start.append(i['season_start'])
    season_end.append(i['season_end'])
    name.append(i['name'])

import pandas as pd
df = pd.DataFrame(list(zip(name, league_id, season, season_start, season_end)), 
                  columns =['name', 'league_id', 'season', 'season_start', 'season_end'])

df.sort_values(by=['name', 'season'])

df.to_csv("../bbdd/spain_leagues.csv", index=False)

primera_division = df['league_id'][df['name']=='Primera Division']
segunda_division = df['league_id'][df['name']=='Segunda Division']

primera_division

segunda_division

for i in primera_division:
    print(i)
