==> notebooks/02_api_get_referees_teams.ipynb <==
import os
os.getcwd()

import pandas as pd 
 
df = pd.read_csv("../bbdd/spain_leagues.csv") 
df.head()

league_id = df['league_id'][(df['name'] == 'Primera Division') | (df['name'] == 'Segunda Division')]

for i in league_id:
    print()

import requests

headers = {'x-rapidapi-key': "28d61524ab5698a1f8edb9e9ed577221"}
teams = []
referee = []

for i in league_id:
    url = "https://server1.api-football.com/fixtures/league/"+str(i)
    response = requests.request("GET", url, headers=headers)
    spain_leagues = response.json()
    
    for k in spain_leagues['api']['fixtures']:
        teams.append(k['homeTeam']['team_name'])
        teams.append(k['awayTeam']['team_name'])
        referee.append(k['referee'])   

teams_list = list(set(teams))
referee_list = list(set(referee))

import pandas as pd

df = pd.DataFrame(teams_list, columns=["teams"])
df.to_excel('../bbdd/teams_list.xlsx', index=False)

df = pd.DataFrame(referee_list, columns=["referees"])
df.to_excel('../bbdd/referees_list.xlsx', index=False, )
