==> notebooks/03_create_dictionaries_json.ipynb <==
import os
os.getcwd()

import pandas as pd

# Load spreadsheet
xl = pd.ExcelFile('../bbdd/dictionary.xlsx')
dictionaries = xl.sheet_names
dictionaries

teams = xl.parse('teams')
tmpzip = zip(teams['ORIGINAL'], teams['FINAL'])
teams_dict = dict(tmpzip)
teams_dict

referees = xl.parse('referees')
tmpzip = zip(referees['ORIGINAL'], referees['FINAL'])
referees_dict = dict(tmpzip)
referees_dict

import json

with open('../bbdd/referees_dict.json', 'w') as outfile:
    json.dump(referees_dict, outfile)
    
with open('../bbdd/teams_dict.json', 'w') as outfile:
    json.dump(teams_dict, outfile)
