==> notebooks/04_clean_historic_ddbb.ipynb <==
import os
os.getcwd()

# read original database. json files / ditionaries will be used to change team names and referees so that
# historical names and api information are the same

import pandas as pd

xl = pd.ExcelFile('../bbdd/FULL_BBDD_ORIGINAL.xls')
df = xl.parse(xl.sheet_names[0])
print(df.head())
df.shape

import json

with open('../bbdd/teams_dict.json') as json_file:
    teams_dict = json.load(json_file)
    
with open('../bbdd/referees_dict.json') as json_file:
    referees_dict = json.load(json_file)

df.columns

df_correct = df.copy()
df_correct['LOCAL'] = [teams_dict[i] for i in df['LOCAL']]
df_correct['VISITANTE'] = [teams_dict[i] for i in df['VISITANTE']]
df_correct['ARBITRO'] = [referees_dict[i] for i in df['ARBITRO']]

df_correct.head()

import numpy as np

df_correct['DIFF'] = df_correct['GOL L'] - df_correct['GOL V']
df_correct['QUINIELA'] = pd.cut(df_correct['DIFF'], bins=[-np.inf, -1, 0, np.inf], labels=['2', 'X', '1'])
df_correct = df_correct.drop(['DIFF'], axis=1)
df_correct.head()

df_correct.shape

df_correct.to_excel("../bbdd/FULL_BBDD_CLEAN.xlsx", index=False)
