==> notebooks/07_scrap_next_quiniela_as.com.ipynb <==
### SCRAP NEXT AVAILABLE QUINIELA

import requests
from bs4 import BeautifulSoup

URL = 'https://resultados.as.com/quiniela/'
page = requests.get(URL)
soup = BeautifulSoup(page.content, 'html.parser')

seccion = soup.find(class_='cont-quiniela')
equipos = seccion.find_all('', class_='nombre-equipo', )

quiniela = [equipo.text for equipo in equipos]

visitor = [equipo for i, equipo in enumerate(quiniela) if i % 2 == 1]
local = [equipo for i, equipo in enumerate(quiniela) if i % 2 == 0]

import pandas as pd
quiniela_df = pd.DataFrame(list(zip(local, visitor)), 
               columns =['LOCAL', 'VISITANTE']) 
quiniela_df

### Check names with dictionary

import json

with open('../bbdd/teams_dict.json') as json_file:
    teams_dict = json.load(json_file)
    
with open('../bbdd/referees_dict.json') as json_file:
    referees_dict = json.load(json_file)

df = quiniela_df.copy()

df['LOCAL'] = [teams_dict[i] for i in df['LOCAL']]
df['VISITANTE'] = [teams_dict[i] for i in df['VISITANTE']]

teams_dict

Next piece of code is meant to use only if new team names are found

teams_dict.update({'Málaga':'Malaga'})

teams_dict['R. Oviedo']

with open('../bbdd/teams_dict.json', 'w') as outfile:
    json.dump(teams_dict, outfile)

Check next Quiniela

df

df.to_excel('../bbdd/QUINIELA_1_CURRENT.xlsx', index=False)
