==> notebooks/10_progress_quinielas.ipynb <==
import os
os.getcwd()

import pandas as pd
import numpy as np
from itertools import product
from random import choices

xl = pd.ExcelFile('../bbdd/QUINIELA_3_PROBS.xlsx')
df = xl.parse(xl.sheet_names[0])[:14]

df

Function to set the maximum streak a result (1, X or 2) can have

def max_streak(df, res):
    probs = []
    for streak in range(14):
        max = 0
        for start in range(14-streak):
            value = 1
            for i in range(start, start+streak+1):
                value = value * df[res][i]
            if value > max:
                max = value
        probs.append(max)
    return(probs)

Function to set the minimum and maximum of 1, X or 2 in each Quiniela

def min_max(df, result, threshold=0.2):
    dictionary_p = {'p'+str(i+1): [df[result][i], 1-df[result][i]] for i in range(14)}
    df_probs = pd.DataFrame([row for row in product(*dictionary_p.values())], columns=dictionary_p.keys())
    probs = df_probs.apply(lambda x: np.prod(x), axis=1)
    
    dictionary_c = {'p'+str(i+1): [1, 0] for i in range(14)}
    df_cases = pd.DataFrame([row for row in product(*dictionary_c.values())], columns=dictionary_c.keys())
    cases = df_cases.apply(lambda x: sum(x), axis=1)
    
    df_cp = pd.DataFrame(zip(cases, probs), columns=(['case','prob']))
    arr_probs = []
    for i in range(14):
        arr_probs.append(sum(df_cp[df_cp['case']==i]['prob']))
    df_cp = pd.DataFrame(zip(range(1,15), arr_probs), columns=(['case','prob']))
    df_cp['prob'] = df_cp['prob']
    df_cp = df_cp.sort_values(by=['prob'], ascending=True)
    df_cp['cumsum'] = np.cumsum(df_cp['prob'])
    return(min(df_cp[df_cp['cumsum']> threshold]['case']), max(df_cp[df_cp['cumsum']> threshold]['case']))  

Calculate maximums, minimums and streaks

min_1, max_1 = min_max(df, '1')
min_x, max_x = min_max(df, 'X')
min_2, max_2 = min_max(df, '2')
streak1 = sum(np.array(max_streak(df, '1')) > 0.1)
streakx = sum(np.array(max_streak(df, 'X')) > 0.1)
streak2 = sum(np.array(max_streak(df, '2')) > 0.1)

bets = 50
df['n1'] = round(bets*df['1'])+1
df['nx'] = round(bets*df['X'])+1
df['n2'] = round(bets*df['2'])+1

df

Function to check how long is a streak in a serie

from itertools import groupby
def eval_streak(serie, res):
    out = [sum(1 for j in group) for j, group in groupby(serie) if j==res]
    if len(out) < 1:
        return(0)
    else:
        return(max(out))

q = pd.DataFrame(np.nan, index=range(14), columns=range(bets))

opcs=0
for quiniela in range(bets):
    correct = False
    while correct == False:
        opcs += 1
        serie = []
        for i in range(14):
            correct = True
            rest1 = df['n1'][i] - sum(q.iloc[i]=='1')
            restx = df['nx'][i] - sum(q.iloc[i]=='X')
            rest2 = df['n2'][i] - sum(q.iloc[i]=='2')
            p1 = rest1 / (rest1 + rest2 + restx)
            p2 = rest2 / (rest1 + rest2 + restx)
            px = restx / (rest1 + rest2 + restx)
            serie.append(choices(['1','X','2'], [p1, px, p2])[0])
            n1 = serie.count('1')
            n2 = serie.count('2')
            nx = serie.count('X')
        correct = False if ((n1 > max_1) | (n1 < min_1) | (n2 > max_2) | (n2 < min_2) | (nx > max_x) | (nx < min_x) | 
                            (eval_streak(serie, '1') > streak1) | (eval_streak(serie, '2') > streak2) | 
                            (eval_streak(serie, 'X') > streakx)) else True
    q[quiniela] = serie

print(opcs, "combinations have been calculated and", bets, "have been selected.")
print(q)

for i in range(1, bets):
    selec = q[i]
    for j in range(i):
        print(i, "=", j, (list(q[i])==list(q[j])))

list(q[1])
