==> notebooks/11_mongodb_conversions.ipynb <==
### Connect to MongoDB

import pymongo

client = pymongo.MongoClient(
    "mongodb+srv://Albert:madannnn@cluster0-wghee.mongodb.net/test?retryWrites=true&w=majority")
db = client.test

try:
    print("MongoDB version is %s" %
            client.server_info()['version'])
except pymongo.errors.OperationFailure as error:
    print(error)
    quit(1)

#accedo o creo bbdd llamada gettingStarted y le asigno nombre 'db'
db = client.gettingStarted

#accedo o creo tabla llamada people dentro de db y le pongo nombre 'people'
people = db.people

#creo 1 documento y lo inserto en people
import datetime
personDocument = {
  "name": { "first": "Alan", "last": "Turing" },
  "birth": datetime.datetime(1912, 6, 23),
  "death": datetime.datetime(1954, 6, 7),
  "contribs": [ "Turing machine", "Turing test", "Turingery" ],
  "views": 1250000
}
people.insert_one(personDocument)

### Date Conversions Tests

from datetime import datetime
timestamp = 1545730073
timestamp = 1576972800
dateobject = datetime.fromtimestamp(timestamp)
dateobject.date()

from datetime import datetime
# current date and time
now = datetime.now()
timestamp = datetime.timestamp(now)
print("timestamp =", timestamp)

### Insert Excel files to MongoDB

import pandas as pd
import numpy as np
import json
from datetime import datetime

xl = pd.ExcelFile('../bbdd/FULL_BBDD_only_results.xlsx')
df = xl.parse(xl.sheet_names[0])

df.tail(3)

records = json.loads(df.T.to_json()).values()

db = client.quiniela
db.onlyresults.drop()

onlyresults = db.onlyresults

onlyresults.insert_many(records)

mongo = list(db.withfeatures.find({}))
df_mongo = pd.DataFrame.from_dict(mongo, orient='columns')
df_mongo.drop(columns=['_id'], inplace=True)
df_mongo.FECHA = df_mongo.FECHA.apply(lambda x: datetime.fromtimestamp(x/1000).date())

df_mongo.head()

### Insert DICT to MongoDB

import pymongo
import pandas as pd
import numpy as np
import json

client = pymongo.MongoClient(
    "mongodb+srv://Albert:madannnn@cluster0-wghee.mongodb.net/test?retryWrites=true&w=majority")
db = client.test

with open('../bbdd/teams_dict.json') as json_file:
    teams_dict = json.load(json_file)

tmp = pd.DataFrame.from_dict(teams_dict, 'index').reset_index()
tmp.columns = ['ORIGINAL','FINAL']
teams_mongo = json.loads(tmp.T.to_json()).values()

#nm = 'Prueba'
inserta = {'ORIGINAL': 'Sporting', 'FINAL': 'Sporting Gijon'}
db = client.quiniela
#db.teamsdict.drop()
teamsdict = db.teamsdict
teamsdict.insert_one(inserta)

db = client.quiniela
db.teamsdict.drop()

teamsdict = db.teamsdict
teamsdict.insert_many(teams_mongo)

Preparo para abrir la BD

mongo = list(db.teamsdict.find({}))

df_mongo = pd.DataFrame.from_dict(mongo, orient='columns')
df_mongo.drop(columns=['_id'], inplace=True)
teams_dict = df_mongo.set_index('ORIGINAL')['FINAL'].to_dict()
